Use int64_t instead of size_t

+++ src/main/Constants.h
@@ -1,7 +1,7 @@
 #pragma once
 
-#include <cstddef>
+#include <cstdint>
 
 static const double INITIAL_SAMPLES_PER_PIXEL = 1;
 static const double INITIAL_VERTICAL_ZOOM = 1;
-static const size_t INITIAL_SAMPLE_OFFSET = 0;
+static const int64_t INITIAL_SAMPLE_OFFSET = 0;
