+++ src/main/gui/SamplePoint.h
@@ -6,7 +6,7 @@
 class SamplePoint : public Component
 {
 private:
-    const size_t sampleIndex;
+    const int64_t sampleIndex;
     const uint8_t channelIndex;
     bool isDragging = false;
     float dragYPos = 0.f;
@@ -14,7 +14,7 @@ private:
     float getSampleValueForYPos(const int16_t y, const uint16_t h, const double v, const uint16_t samplePointSize);
 
 public:
-    SamplePoint(CupuacuState*, const uint8_t channelIndexToUse, const size_t sampleIndexToUse);
+    SamplePoint(CupuacuState*, const uint8_t channelIndexToUse, const int64_t sampleIndexToUse);
 
     uint64_t getSampleIndex() const;
     float getSampleValue() const;
